optimize: Transaction timeout judgment is performed on the TM side and TC side

+++ common/src/main/java/io/seata/common/DefaultValues.java
@@ -90,7 +90,7 @@ public interface DefaultValues {
     int DEFAULT_TM_ROLLBACK_RETRY_COUNT = 5;
     int DEFAULT_GLOBAL_TRANSACTION_TIMEOUT = 60000;
 
-    long DEFAULT_TM_LOSS_TIME = 1;
+    float DEFAULT_TM_LOSS_TIME = 1L;
 
     String DEFAULT_TX_GROUP = "default_tx_group";
     @Deprecated
